change(): Move LoginRequest.php to Auth namespace
